v3.7

+++ shinelink/本地调试/一键诊断/YDChart/CustomProgress.h
@@ -13,5 +13,5 @@
 @property(nonatomic, retain)UIImageView *leftimg;
 @property(nonatomic, retain)UILabel *presentlab;
 @property(nonatomic)float maxValue;
--(void)setPresent:(int)present;
+-(void)setPresent:(float)present;
 @end
